Trie (#21)
* refactor assert

* implement trie list

* start tests

* run tests

* enable optmization

* include stdint

* format

* fix bug

* improve performance

* improve tests

+++ lib/list/head/debug.h
@@ -7,9 +7,9 @@
 
 #define LH(POINTER) ((list_head_p)(POINTER))
 
-list_head_p clu_list_head_create_immed(int n, ...);
+list_head_p clu_list_head_create_immed(uint64_t n, ...);
 
-bool clu_list_head_immed(list_head_p lh, int n, ...);
+bool clu_list_head_immed(list_head_p lh, uint64_t n, ...);
 
 #endif
